+++ R/lifecycle_admiral.R
@@ -1,0 +1,35 @@
+#' Deprecation with Soft Message
+#'
+#' Wrapper around `lifecycle::deprecate_soft()`.
+#' See documentation for `lifecycle::deprecate_soft()` for argument use.
+#'
+#' @param ... See documentation for `lifecycle::deprecate_soft()` for argument use.
+#'
+#' @return `NULL`, invisibly.
+#'
+#' @examples
+#' # A Phase 1 deprecated function with custom bulleted list:
+#' deprecate_inform(
+#'   when = "1.0.0",
+#'   what = "foo()",
+#'   details = c(
+#'     #     'x = "This message will turn into a warning with release of x.y.z",
+#'     i = "See admiral's deprecation guidance:
+#' https://pharmaverse.github.io/admiraldev/dev/articles/programming_strategy.html#deprecation"
+#'   )
+#' )
+#'
+#' @keywords messages
+#' @family messages
+#'
+#'
+#' @export
+deprecate_inform <- function(...) {
+  tryCatch(
+    lifecycle::deprecate_soft(...),
+    warning = \(w) {
+      message(conditionMessage(w))
+      tryInvokeRestart("muffleWarning")
+    }
+  )
+}