+++ tests/testthat/test-lifecycle_admiral.R
@@ -1,0 +1,43 @@
+## Test 1:  Message is sent to user ----
+test_that("lifecycle_admiral Test 1:  Message is sent to user", {
+  example_fun <- function(dataset) {
+    deprecate_inform("1.0.0", "example_fun()", "example_fun2()")
+    assert_data_frame(dataset, required_vars = exprs(STUDYID, USUBJID))
+  }
+
+  data <- dplyr::tribble(
+    ~STUDYID, ~USUBJID,
+    "xyz",    123,
+    "xyz",    456
+  )
+
+  expect_snapshot({
+    example_fun(data)
+    example_fun(data)
+  })
+})
+
+# Test 2:  Nicer message is sent to user ----
+test_that("lifecycle_admiral Test 2:  Nicer message is sent to user", {
+  example_fun <- function(dataset) {
+    deprecate_inform("1.0.0", "example_fun()", "example_fun2()",
+      details = c(
+        x = "This message will turn into a warning with release of 1.1.0",
+        i = "See admiral's deprecation guidance:
+                       https://pharmaverse.github.io/admiraldev/dev/articles/programming_strategy.html#deprecation" # nolint
+      )
+    )
+    assert_data_frame(dataset, required_vars = exprs(STUDYID, USUBJID))
+  }
+
+  data <- dplyr::tribble(
+    ~STUDYID, ~USUBJID,
+    "xyz",    123,
+    "xyz",    456
+  )
+
+  expect_snapshot({
+    example_fun(data)
+    example_fun(data)
+  })
+})