+++ NEWS.md
@@ -1,5 +1,10 @@
 # admiraldev (development version)
 
+## New Features
+
+- New Function `deprecate_inform()` is a wrapper around 
+  `lifecycle::deprecate_soft()` to allow for more control over messaging. (#466)
+
 ## Updates of Existing Functions
 
 - Objects passed to `assert_list_element()` via the `...` argument can now be
@@ -27,6 +32,7 @@ check whether the vector has a specific length. (#473)
   - Removed at v1.0.0 `assert_function_param()`
 
 ## Documentation
+  - Deprecation Strategy updated for the long haul! (#466)
 
 ## Other